+++ plugins/htmlwriter/plugin.js
@@ -16,7 +16,7 @@ CKEDITOR.plugins.add( 'htmlwriter', {
 } );
 
 /**
- * Class used to write HTML data.
+ * The class used to write HTML data.
  *
  *		var writer = new CKEDITOR.htmlWriter();
  *		writer.openTag( 'p' );
@@ -33,7 +33,7 @@ CKEDITOR.htmlWriter = CKEDITOR.tools.createClass( {
 	base: CKEDITOR.htmlParser.basicWriter,
 
 	/**
-	 * Creates a htmlWriter class instance.
+	 * Creates an `htmlWriter` class instance.
 	 *
 	 * @constructor
 	 */
@@ -42,7 +42,7 @@ CKEDITOR.htmlWriter = CKEDITOR.tools.createClass( {
 		this.base();
 
 		/**
-		 * The characters to be used for each identation step.
+		 * The characters to be used for each indentation step.
 		 *
 		 *		// Use tab for indentation.
 		 *		editorInstance.dataProcessor.writer.indentationChars = '\t';
@@ -105,7 +105,7 @@ CKEDITOR.htmlWriter = CKEDITOR.tools.createClass( {
 
 	proto: {
 		/**
-		 * Writes the tag opening part for a opener tag.
+		 * Writes the tag opening part for an opener tag.
 		 *
 		 *		// Writes '<p'.
 		 *		writer.openTag( 'p', { class : 'MyClass', id : 'MyId' } );
@@ -134,7 +134,7 @@ CKEDITOR.htmlWriter = CKEDITOR.tools.createClass( {
 		},
 
 		/**
-		 * Writes the tag closing part for a opener tag.
+		 * Writes the tag closing part for an opener tag.
 		 *
 		 *		// Writes '>'.
 		 *		writer.openTagClose( 'p', false );
@@ -265,7 +265,7 @@ CKEDITOR.htmlWriter = CKEDITOR.tools.createClass( {
 		},
 
 		/**
-		 * Writes the current indentation chars. It uses the {@link #indentationChars}
+		 * Writes the current indentation character. It uses the {@link #indentationChars}
 		 * property, repeating it for the current indentation steps.
 		 *
 		 *		// Writes '\t' (e.g.).
@@ -292,21 +292,21 @@ CKEDITOR.htmlWriter = CKEDITOR.tools.createClass( {
 		},
 
 		/**
-		 * Sets formatting rules for a give element. The possible rules are:
+		 * Sets formatting rules for a given element. Possible rules are:
 		 *
-		 * * `indent`: indent the element contents.
-		 * * `breakBeforeOpen`: break line before the opener tag for this element.
-		 * * `breakAfterOpen`: break line after the opener tag for this element.
-		 * * `breakBeforeClose`: break line before the closer tag for this element.
-		 * * `breakAfterClose`: break line after the closer tag for this element.
+		 * * `indent` &ndash; indent the element content.
+		 * * `breakBeforeOpen` &ndash; break line before the opener tag for this element.
+		 * * `breakAfterOpen` &ndash; break line after the opener tag for this element.
+		 * * `breakBeforeClose` &ndash; break line before the closer tag for this element.
+		 * * `breakAfterClose` &ndash; break line after the closer tag for this element.
 		 *
-		 * All rules default to `false`. Each call to the function overrides
-		 * already present rules, leaving the undefined untouched.
+		 * All rules default to `false`. Each function call overrides rules that are
+		 * already present, leaving the undefined ones untouched.
 		 *
 		 * By default, all elements available in the {@link CKEDITOR.dtd#$block},
-		 * {@link CKEDITOR.dtd#$listItem} and {@link CKEDITOR.dtd#$tableContent}
+		 * {@link CKEDITOR.dtd#$listItem}, and {@link CKEDITOR.dtd#$tableContent}
 		 * lists have all the above rules set to `true`. Additionaly, the `<br>`
-		 * element has the `breakAfterOpen` set to `true`.
+		 * element has the `breakAfterOpen` rule set to `true`.
 		 *
 		 *		// Break line before and after "img" tags.
 		 *		writer.setRules( 'img', {
@@ -317,7 +317,7 @@ CKEDITOR.htmlWriter = CKEDITOR.tools.createClass( {
 		 *		// Reset the rules for the "h1" tag.
 		 *		writer.setRules( 'h1', {} );
 		 *
-		 * @param {String} tagName The element name to which set the rules.
+		 * @param {String} tagName The name of the element for which the rules are set.
 		 * @param {Object} rules An object containing the element rules.
 		 */
 		setRules: function( tagName, rules ) {
@@ -332,8 +332,8 @@ CKEDITOR.htmlWriter = CKEDITOR.tools.createClass( {
 } );
 
 /**
- * Whether to force using `'&'` instead of `'&amp;'` in elements attributes
- * values, it's not recommended to change this setting for compliance with the
+ * Whether to force using `'&'` instead of `'&amp;'` in element attributes
+ * values. It is not recommended to change this setting for compliance with the
  * W3C XHTML 1.0 standards ([C.12, XHTML 1.0](http://www.w3.org/TR/xhtml1/#C_12)).
  *
  *		// Use `'&'` instead of `'&amp;'`
@@ -344,9 +344,9 @@ CKEDITOR.htmlWriter = CKEDITOR.tools.createClass( {
  */
 
 /**
- * The characters to be used for indenting the HTML produced by the editor.
- * Using characters different than `' '` (space) and `'\t'` (tab) is definitely
- * a bad idea as it'll mess the code.
+ * The characters to be used for indenting HTML output produced by the editor.
+ * Using characters different from `' '` (space) and `'\t'` (tab) is not recommended
+ * as it will mess the code.
  *
  *		// No indentation.
  *		CKEDITOR.config.dataIndentationChars = '';